Restore EntityAttribute fields in EntityAttributes
This just remaps the new yarn intermediary names for the EntityAttribute fields in EntityAttributes to the names they used to have. I'm mapping these because I've been struggling to port a couple mods to the latest snapshot, only to find that the fields I need aren't even mapped. I don't really know how to use Github, and I don't know if I'm respecting the policies regarding other mapping projects correctly. I just needed this to get the projects I'm porting to compile, so.